chore: refresh decorations on scroll
Problem: decorations are not displayed when the active
editor was scrolled using a mouse and/or cursor position
is out of the visible range.
Solution: display decorations even if the cursor position
is out of view. Pick the middle line as a center of focus
rather than the line of the cursor anchor in case it is
not visible.

Method `showDecorations` is now wrapped in a debounce
with 80ms timeout.